% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{replace_email}
\alias{replace_email}
\title{Function to mask emails}
\usage{
replace_email(x)
}
\arguments{
\item{x}{character string that should be masked}
}
\value{
string with email masked as \verb{[email]}
}
\description{
\code{replace_email()} takes an input string and looks for common email structure. If an email is found ,it replaces the email
with \verb{[email]}
}
\examples{
replace_email("my test string")
}
